Remove live region markup (#9)
* Remove live region markup
I would remove this ARIA markup. It causes a conflict when using the date pickers with screen readers. The whole contents of `<ul id='steps'>` are announced before the selected day/month/year, which makes using the pickers quite an inefficient experience. I apprecieate the intent to make screen reader users aware of the changes in the different milestones as you move through the calendar, but I think a better experience is for the user to first select the date they want and then to review the changes in the other milestones by themselves using regular screen reader navigation commands.

* Now save changes and add them to git